# Refactor the block-editor parents state to use maps instead of objects.