Similar to #46204 #46221 and #46146

Following a similar approach to the previous PRs, I tried to see how refactoring the parents state from plain objects to maps could impact the performance. Here are the results I got:

Before:

<img width="515" alt="Screenshot 2022-12-01 at 12 32 39 PM" src="https://user-images.githubusercontent.com/272444/205042881-40fe9321-4085-4048-9cf3-6f5536ab5f64.png">

After:

<img width="515" alt="Screenshot 2022-12-01 at 12 31 51 PM" src="https://user-images.githubusercontent.com/272444/205042775-486fe568-adda-441a-9c52-ad2f0ab31a37.png">

As you can see:

 - The "reset blocks" (loading) metric is stable improved from 8s to 6.3s
 - The "hit enter to create a new block" metric improved from 267ms to 190ms
 - The "update blocks" metric degraded from 69ms to 80ms (I'll see if I can debug the degradation here to see if we can do something about it)