@jasmussen 2022-12-12
I think this PR had an unintended side effect of making it impossible to select the inner blocks of a contentOnly locked pattern. Here's a sample pattern:

```
<!-- wp:group {"templateLock":"contentOnly","layout":{"type":"constrained"}} -->
<div class="wp-block-group"><!-- wp:columns -->
<div class="wp-block-columns"><!-- wp:column -->
<div class="wp-block-column"><!-- wp:image {"id":525,"sizeSlug":"large","linkDestination":"none"} -->
<figure class="wp-block-image size-large"><img src="https://s.w.org/patterns/files/2021/06/Iris-793x1024.jpg" alt="" class="wp-image-525"/></figure>
<!-- /wp:image -->

<!-- wp:paragraph {"style":{"typography":{"fontSize":"14px"}}} -->
<p style="font-size:14px"><strong>White Irises</strong><br>Ogawa Kazumasa</p>
<!-- /wp:paragraph --></div>
<!-- /wp:column -->

<!-- wp:column -->
<div class="wp-block-column"><!-- wp:spacer -->
<div style="height:100px" aria-hidden="true" class="wp-block-spacer"></div>
<!-- /wp:spacer -->

<!-- wp:paragraph {"style":{"typography":{"fontSize":"14px"}}} -->
<p style="font-size:14px"><strong>Cherry Blossom</strong><br>Ogawa Kazumasa</p>
<!-- /wp:paragraph -->

<!-- wp:image {"id":524,"sizeSlug":"large","linkDestination":"none"} -->
<figure class="wp-block-image size-large"><img src="https://s.w.org/patterns/files/2021/06/Cherry-Blossom-707x1024.jpg" alt="" class="wp-image-524"/></figure>
<!-- /wp:image --></div>
<!-- /wp:column --></div>
<!-- /wp:columns --></div>
<!-- /wp:group -->
```

I found this PR as the "breaking" one purely by shotgun debugging, checking out [commits](https://github.com/WordPress/gutenberg/commits/trunk?before=7dda7859adfa06adbe0da026f8f1253c2b001847+136&branch=trunk&qualified_name=refs%2Fheads%2Ftrunk), and specifically the behavior broke between these two commits:

<img width="1226" alt="Screenshot 2022-12-12 at 14 42 54" src="https://user-images.githubusercontent.com/1204802/207060300-a2a7f959-9ca7-48af-9dfe-13b4d41c0a78.png">

Before the topmost of those two commits, it worked like this:

![before](https://user-images.githubusercontent.com/1204802/207060458-3cd710cd-928f-47b4-ad3d-f34e163f4026.gif)

After that top commit, it's now brokenb, like this:

![after](https://user-images.githubusercontent.com/1204802/207060665-2944e7ba-c379-4c1f-912c-d7b6261122e1.gif)

CC: @youknowriad @jorgefilipecosta