@youknowriad 2022-12-13
Thanks for catching that, fix in https://github.com/WordPress/gutenberg/pull/46494