@dcalhoun 2023-04-14
Hey @youknowriad. 👋🏻 I reviewed this PR and the related ones listed in the description while exploring typing performance in the native mobile editor. At a high-level, do you consider the effort to use `Map` over `Object` complete or do you perceive additional portions of the reducer that might benefit from this type of refactor? Thanks! 🙇🏻