@youknowriad 2023-04-17
Hey @dcalhoun yeah at the moment, I think this is "done" from my point of view. I do think there's more we can do in the reducer but I remember doing more of these changes in some of parts of the reducer and not measuring the perf gains like I did for the merged PRs. So yeah, no changes coming in this area from me, but let's keep an eye for potential improvements.