@jsnajdr 2023-04-17
Some time ago I created a draft PR that migrates the `blockListSettings` state from object to a map. And also updates the `updateBlockListSettings` action to accept an array of updates to be performed as a batch. But I don't thing there were measurable performance benefits just from this change. One day I might revisit that PR and get it merged.